+++ src/ConifgWebView.ts
@@ -1,10 +1,15 @@
 import * as vscode from "vscode"
 import { Uri, Webview, WebviewView, WebviewViewProvider } from "vscode"
-import { SUPPORTED_MODELS_DETAILS, SystemInfo } from "./resourceManager"
+import {
+    NO_WORKSPACE_OPEN,
+    SUPPORTED_MODELS_DETAILS,
+    SystemInfo,
+} from "./resourceManager"
+import { updateLuaLibraryConfigurations } from "./workspaceManager"
 
 export class ConfigWebView implements WebviewViewProvider {
     public static readonly viewType = "systemConfigurations"
-    constructor(private readonly _extensionUri: Uri) {}
+    constructor(private readonly _extensionUri: Uri) { }
     resolveWebviewView(
         webviewView: vscode.WebviewView,
         // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -26,6 +31,28 @@ export class ConfigWebView implements WebviewViewProvider {
         // Sets up an event listener to listen for messages passed from the webview view context
         // and executes code based on the message that is recieved
         this._setWebviewMessageListener(webviewView)
+
+        vscode.commands.registerCommand(
+            "systemConfigurations.addSystem",
+            () => {
+                if (!vscode.workspace.workspaceFolders) {
+                    vscode.window.showInformationMessage(`${NO_WORKSPACE_OPEN}`)
+                    return
+                }
+                webviewView.webview.postMessage({
+                    command: "supportedModels",
+                    payload: JSON.stringify(SUPPORTED_MODELS_DETAILS),
+                })
+            },
+        )
+        // Register a callback for configuration changes
+        vscode.workspace.onDidChangeConfiguration(async (event) => {
+            if (event.affectsConfiguration("tsp.tspLinkSystemConfigurations")) {
+                await this.getSystemName()
+                this.reloadUi(webviewView)
+                await updateLuaLibraryConfigurations()
+            }
+        })
     }
     private _getWebviewContent(webview: Webview) {
         if (!vscode.workspace.workspaceFolders) {
@@ -33,7 +60,7 @@ export class ConfigWebView implements WebviewViewProvider {
             <html lang="en">
             <head></head>
             <body>
-            <h1>No workspace has been open, please open an workspace and reload this UI</h1>
+            <h1>${NO_WORKSPACE_OPEN}</h1>
             </body>
             </html>`
         } else {
@@ -80,32 +107,13 @@ export class ConfigWebView implements WebviewViewProvider {
             // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
             switch (message.command) {
                 case "getInitialSystems": {
-                    const savedSystems: SystemInfo[] =
-                        vscode.workspace
-                            .getConfiguration("tsp")
-                            .get("tspLinkSystemConfigurations") ?? []
-
-                    webviewView.webview.postMessage({
-                        command: "systems",
-                        payload: JSON.stringify({
-                            systemInfo: savedSystems,
-                            supportedModels: SUPPORTED_MODELS_DETAILS,
-                        }),
-                    })
-                    break
-                }
-                case "getSupportedModels": {
-                    webviewView.webview.postMessage({
-                        command: "supportedModels",
-                        payload: JSON.stringify(SUPPORTED_MODELS_DETAILS),
-                    })
+                    this.reloadUi(webviewView)
                     break
                 }
                 case "add":
                     {
                         // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
                         const newSystemDetails = message.data as SystemInfo
-                        console.log(newSystemDetails)
                         const originalSystemInfo: SystemInfo[] =
                             vscode.workspace
                                 .getConfiguration("tsp")
@@ -220,10 +228,50 @@ export class ConfigWebView implements WebviewViewProvider {
         })
     }
 
+    private reloadUi(webviewView: vscode.WebviewView) {
+        const savedSystems: SystemInfo[] =
+            vscode.workspace
+                .getConfiguration("tsp")
+                .get("tspLinkSystemConfigurations") ?? []
+
+        webviewView.webview.postMessage({
+            command: "systems",
+            payload: JSON.stringify({
+                systemInfo: savedSystems,
+                supportedModels: SUPPORTED_MODELS_DETAILS,
+            }),
+        })
+    }
+    private async getSystemName() {
+        const existingSystems: SystemInfo[] =
+            vscode.workspace
+                .getConfiguration("tsp")
+                .get("tspLinkSystemConfigurations") ?? []
+
+        const systemWithEmptyName = existingSystems.find(
+            (system) => !system.name,
+        )
+        if (systemWithEmptyName) {
+            const options: vscode.InputBoxOptions = {
+                prompt: "Enter new system name",
+            }
+            const name = await vscode.window.showInputBox(options)
+            if (name)
+                systemWithEmptyName.name = name
+            await vscode.workspace
+                .getConfiguration("tsp")
+                .update(
+                    "tspLinkSystemConfigurations",
+                    existingSystems,
+                    false,
+                )
+        }
+    }
+
     private getNonce() {
         let text = ""
         const possible =
-            "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
+        "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789"
         for (let i = 0; i < 32; i++) {
             text += possible.charAt(Math.floor(Math.random() * possible.length))
         }